buffer jump input + fix OnJump event
i was looking at the player script, and apparently the part that makes the player jump is called twice every time you input jump (because the frame after the jump you are still quite close to the ground). Making it so it's only called once every jump makes the jump height slightly off from calculation. So i'm leaving it to be called twice, but made sure that OnJump only gets triggered once.